🐛 Swagger https 적용
스웨거에서 API를 호출할 때 http와 https 중에 선택할 수 있습니다.

개발용 url과 배포 서버용 url을 구분했습니다.
