--- src/main/java/dnd/donworry/SwaggerConfig.java
@@ -1,25 +1,0 @@
-package dnd.donworry;
-
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI()
-                .components(new Components())
-                .info(apiInfo());
-    }
-
-    private Info apiInfo() {
-        return new Info()
-                .title("API Test") // API의 제목
-                .description("DonWorry Swagger UI") // API에 대한 설명
-                .version("1.0.0"); // API의 버전
-    }
-}
